+++ src/basic/pages/CartPage.tsx
@@ -1,0 +1,434 @@
+import { CartItem, Coupon } from "../../types";
+import { Product } from "../features/product/types";
+import { ProductList } from "../features/product/components/ProductList";
+import {
+  calculateDiscountedPrice,
+  calculateDiscountedAmount,
+  formatPrice,
+} from "../shared/libs/price";
+import {
+  getProductStockStatus,
+  getRemainingStock,
+} from "../features/product/libs";
+import { useCallback } from "react";
+
+interface ProductWithUI extends Product {
+  description?: string;
+  isRecommended?: boolean;
+}
+
+interface CartPageProps {
+  products: ProductWithUI[];
+  filteredProducts: ProductWithUI[];
+  searchValue: string;
+  cart: CartItem[];
+  setCart: (cart: CartItem[]) => void;
+  coupons: Coupon[];
+  selectedCoupon: Coupon | null;
+  setSelectedCoupon: (coupon: Coupon | null) => void;
+  addNotification: (
+    message: string,
+    type?: "error" | "success" | "warning"
+  ) => void;
+}
+
+export function CartPage({
+  products,
+  filteredProducts,
+  searchValue,
+  cart,
+  setCart,
+  coupons,
+  selectedCoupon,
+  setSelectedCoupon,
+  addNotification,
+}: CartPageProps) {
+  const displayPrice = (product: Product) => {
+    const stockStatus = getProductStockStatus({ product, cartQuantity: 0 });
+    if (stockStatus) return stockStatus;
+
+    const formattedPrice = formatPrice(product.price);
+    return `₩${formattedPrice}`;
+  };
+
+  const getProductRemainingStock = (product: Product): number => {
+    const cartItem = cart.find((item) => item.product.id === product.id);
+    const cartQuantity = cartItem?.quantity || 0;
+    return getRemainingStock(product, cartQuantity);
+  };
+
+  const getMaxApplicableDiscount = (item: CartItem): number => {
+    const { discounts } = item.product;
+    const { quantity } = item;
+
+    const baseDiscount = discounts.reduce((maxDiscount, discount) => {
+      return quantity >= discount.quantity && discount.rate > maxDiscount
+        ? discount.rate
+        : maxDiscount;
+    }, 0);
+
+    const hasBulkPurchase = cart.some((cartItem) => cartItem.quantity >= 10);
+    if (hasBulkPurchase) {
+      return Math.min(baseDiscount + 0.05, 0.5); // 대량 구매 시 추가 5% 할인
+    }
+
+    return baseDiscount;
+  };
+
+  const calculateItemTotal = (item: CartItem): number => {
+    const { price } = item.product;
+    const { quantity } = item;
+    const discount = getMaxApplicableDiscount(item);
+
+    const discountedPrice = calculateDiscountedPrice(price, discount * 100);
+    return Math.round(discountedPrice * quantity);
+  };
+
+  const calculateCartTotal = (): {
+    totalBeforeDiscount: number;
+    totalAfterDiscount: number;
+  } => {
+    let totalBeforeDiscount = 0;
+    let totalAfterDiscount = 0;
+
+    cart.forEach((item) => {
+      const itemPrice = item.product.price * item.quantity;
+      totalBeforeDiscount += itemPrice;
+      totalAfterDiscount += calculateItemTotal(item);
+    });
+
+    if (selectedCoupon) {
+      if (selectedCoupon.discountType === "amount") {
+        totalAfterDiscount = calculateDiscountedAmount(
+          totalAfterDiscount,
+          selectedCoupon.discountValue
+        );
+      } else {
+        totalAfterDiscount = Math.round(
+          calculateDiscountedPrice(
+            totalAfterDiscount,
+            selectedCoupon.discountValue
+          )
+        );
+      }
+    }
+
+    return {
+      totalBeforeDiscount: Math.round(totalBeforeDiscount),
+      totalAfterDiscount: Math.round(totalAfterDiscount),
+    };
+  };
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getProductRemainingStock(product);
+      if (remainingStock <= 0) {
+        addNotification("재고가 부족합니다!", "error");
+        return;
+      }
+
+      setCart(
+        cart
+          .map((item) => {
+            if (item.product.id === product.id) {
+              const newQuantity = item.quantity + 1;
+              if (newQuantity > product.stock) {
+                addNotification(
+                  `재고는 ${product.stock}개까지만 있습니다.`,
+                  "error"
+                );
+                return item;
+              }
+              return { ...item, quantity: newQuantity };
+            }
+            return item;
+          })
+          .concat(
+            cart.find((item) => item.product.id === product.id)
+              ? []
+              : [{ product, quantity: 1 }]
+          )
+      );
+
+      addNotification("장바구니에 담았습니다", "success");
+    },
+    [cart, setCart, addNotification]
+  );
+
+  const removeFromCart = useCallback(
+    (productId: string) => {
+      setCart(cart.filter((item) => item.product.id !== productId));
+    },
+    [cart, setCart]
+  );
+
+  const updateQuantity = useCallback(
+    (productId: string, newQuantity: number) => {
+      if (newQuantity <= 0) {
+        removeFromCart(productId);
+        return;
+      }
+
+      const product = products.find((p) => p.id === productId);
+      if (!product) return;
+
+      const maxStock = product.stock;
+      if (newQuantity > maxStock) {
+        addNotification(`재고는 ${maxStock}개까지만 있습니다.`, "error");
+        return;
+      }
+
+      setCart(
+        cart.map((item) =>
+          item.product.id === productId
+            ? { ...item, quantity: newQuantity }
+            : item
+        )
+      );
+    },
+    [cart, setCart, products, addNotification, removeFromCart]
+  );
+
+  const applyCoupon = useCallback(
+    (coupon: Coupon) => {
+      const currentTotal = calculateCartTotal().totalAfterDiscount;
+
+      if (currentTotal < 10000 && coupon.discountType === "percentage") {
+        addNotification(
+          "percentage 쿠폰은 10,000원 이상 구매 시 사용 가능합니다.",
+          "error"
+        );
+        return;
+      }
+
+      setSelectedCoupon(coupon);
+      addNotification("쿠폰이 적용되었습니다.", "success");
+    },
+    [addNotification, setSelectedCoupon]
+  );
+
+  const completeOrder = useCallback(() => {
+    const orderNumber = `ORD-${Date.now()}`;
+    addNotification(
+      `주문이 완료되었습니다. 주문번호: ${orderNumber}`,
+      "success"
+    );
+    setCart([]);
+    setSelectedCoupon(null);
+  }, [addNotification, setCart, setSelectedCoupon]);
+
+  const totals = calculateCartTotal();
+
+  return (
+    <div className="grid grid-cols-1 lg:grid-cols-4 gap-6">
+      <div className="lg:col-span-3">
+        {/* 상품 목록 */}
+        <ProductList
+          products={products}
+          filteredProducts={filteredProducts}
+          searchValue={searchValue}
+          getProductRemainingStock={getProductRemainingStock}
+          displayPrice={displayPrice}
+          onAddToCart={addToCart}
+        />
+      </div>
+
+      <div className="lg:col-span-1">
+        <div className="sticky top-24 space-y-4">
+          <section className="bg-white rounded-lg border border-gray-200 p-4">
+            <h2 className="text-lg font-semibold mb-4 flex items-center">
+              <svg
+                className="w-5 h-5 mr-2"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M16 11V7a4 4 0 00-8 0v4M5 9h14l1 12H4L5 9z"
+                />
+              </svg>
+              장바구니
+            </h2>
+            {cart.length === 0 ? (
+              <div className="text-center py-8">
+                <svg
+                  className="w-16 h-16 text-gray-300 mx-auto mb-4"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={1}
+                    d="M16 11V7a4 4 0 00-8 0v4M5 9h14l1 12H4L5 9z"
+                  />
+                </svg>
+                <p className="text-gray-500 text-sm">장바구니가 비어있습니다</p>
+              </div>
+            ) : (
+              <div className="space-y-3">
+                {cart.map((item) => {
+                  const itemTotal = calculateItemTotal(item);
+                  const originalPrice = item.product.price * item.quantity;
+                  const hasDiscount = itemTotal < originalPrice;
+                  const discountRate = hasDiscount
+                    ? Math.round((1 - itemTotal / originalPrice) * 100)
+                    : 0;
+
+                  return (
+                    <div
+                      key={item.product.id}
+                      className="border-b pb-3 last:border-b-0"
+                    >
+                      <div className="flex justify-between items-start mb-2">
+                        <h4 className="text-sm font-medium text-gray-900 flex-1">
+                          {item.product.name}
+                        </h4>
+                        <button
+                          onClick={() => removeFromCart(item.product.id)}
+                          className="text-gray-400 hover:text-red-500 ml-2"
+                        >
+                          <svg
+                            className="w-4 h-4"
+                            fill="none"
+                            stroke="currentColor"
+                            viewBox="0 0 24 24"
+                          >
+                            <path
+                              strokeLinecap="round"
+                              strokeLinejoin="round"
+                              strokeWidth={2}
+                              d="M6 18L18 6M6 6l12 12"
+                            />
+                          </svg>
+                        </button>
+                      </div>
+                      <div className="flex items-center justify-between">
+                        <div className="flex items-center">
+                          <button
+                            onClick={() =>
+                              updateQuantity(item.product.id, item.quantity - 1)
+                            }
+                            className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
+                          >
+                            <span className="text-xs">−</span>
+                          </button>
+                          <span className="mx-3 text-sm font-medium w-8 text-center">
+                            {item.quantity}
+                          </span>
+                          <button
+                            onClick={() =>
+                              updateQuantity(item.product.id, item.quantity + 1)
+                            }
+                            className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
+                          >
+                            <span className="text-xs">+</span>
+                          </button>
+                        </div>
+                        <div className="text-right">
+                          {hasDiscount && (
+                            <span className="text-xs text-red-500 font-medium block">
+                              -{discountRate}%
+                            </span>
+                          )}
+                          <p className="text-sm font-medium text-gray-900">
+                            {Math.round(itemTotal).toLocaleString()}원
+                          </p>
+                        </div>
+                      </div>
+                    </div>
+                  );
+                })}
+              </div>
+            )}
+          </section>
+
+          {cart.length > 0 && (
+            <>
+              <section className="bg-white rounded-lg border border-gray-200 p-4">
+                <div className="flex items-center justify-between mb-3">
+                  <h3 className="text-sm font-semibold text-gray-700">
+                    쿠폰 할인
+                  </h3>
+                  <button className="text-xs text-blue-600 hover:underline">
+                    쿠폰 등록
+                  </button>
+                </div>
+                {coupons.length > 0 && (
+                  <select
+                    className="w-full text-sm border border-gray-300 rounded px-3 py-2 focus:outline-none focus:border-blue-500"
+                    value={selectedCoupon?.code || ""}
+                    onChange={(e) => {
+                      const coupon = coupons.find(
+                        (c) => c.code === e.target.value
+                      );
+                      if (coupon) applyCoupon(coupon);
+                      else setSelectedCoupon(null);
+                    }}
+                  >
+                    <option value="">쿠폰 선택</option>
+                    {coupons.map((coupon) => (
+                      <option key={coupon.code} value={coupon.code}>
+                        {coupon.name} (
+                        {coupon.discountType === "amount"
+                          ? `${coupon.discountValue.toLocaleString()}원`
+                          : `${coupon.discountValue}%`}
+                        )
+                      </option>
+                    ))}
+                  </select>
+                )}
+              </section>
+
+              <section className="bg-white rounded-lg border border-gray-200 p-4">
+                <h3 className="text-lg font-semibold mb-4">결제 정보</h3>
+                <div className="space-y-2 text-sm">
+                  <div className="flex justify-between">
+                    <span className="text-gray-600">상품 금액</span>
+                    <span className="font-medium">
+                      {totals.totalBeforeDiscount.toLocaleString()}원
+                    </span>
+                  </div>
+                  {totals.totalBeforeDiscount - totals.totalAfterDiscount >
+                    0 && (
+                    <div className="flex justify-between text-red-500">
+                      <span>할인 금액</span>
+                      <span>
+                        -
+                        {(
+                          totals.totalBeforeDiscount - totals.totalAfterDiscount
+                        ).toLocaleString()}
+                        원
+                      </span>
+                    </div>
+                  )}
+                  <div className="flex justify-between py-2 border-t border-gray-200">
+                    <span className="font-semibold">결제 예정 금액</span>
+                    <span className="font-bold text-lg text-gray-900">
+                      {totals.totalAfterDiscount.toLocaleString()}원
+                    </span>
+                  </div>
+                </div>
+
+                <button
+                  onClick={completeOrder}
+                  className="w-full mt-4 py-3 bg-yellow-400 text-gray-900 rounded-md font-medium hover:bg-yellow-500 transition-colors"
+                >
+                  {totals.totalAfterDiscount.toLocaleString()}원 결제하기
+                </button>
+
+                <div className="mt-3 text-xs text-gray-500 text-center">
+                  <p>* 실제 결제는 이루어지지 않습니다</p>
+                </div>
+              </section>
+            </>
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
